requirements.txt
fastapi==0.75.1
pydantic==1.9.0
tensorflow-gpu==2.3.0
opencv-python~=4.4.0.46
opencv-contrib-python~=4.5.5
numpy~=1.18.5
requests==2.23.0
python-dotenv==0.20.0